+++ src/core/game.ts
@@ -102,7 +102,32 @@ namespace Lightning {
             this._physicsWorldBounds.type = Box2D.Dynamics.b2Body.b2_dynamicBody;
 
             let body = Box2D.Dynamics.b2Body;
-            
+        }
+
+        generateTexture(...params):Texture | Array<Texture> {
+            let textures:Array<Texture> = [];
+            for(let i of params) {
+                textures.push(this._renderer.generateTexture(i));
+            }
+            if(textures.length === 1) {
+                return textures[0];
+            } else {
+                return textures;
+            }
+        }
+
+        texture(...params):any {
+            let t:Texture | Array<Texture> = [];
+            if(params.length > 1) {
+                console.log('get multiple textures');
+                for(let i of params) {
+                    t.push(Texture.from(i));
+                }
+            } else {
+                console.log('get single texture')
+                t = Texture.from(params[0]);
+            }
+            return t;
         }
 
         public set backgroundColor(val:number) {