+++ src/display/texture.ts
@@ -1,7 +1,7 @@
 /// <reference path="./../reference.d.ts" />
 
 namespace Lightning {
-    export default class Texture extends PIXI.Texture {
+    export class Texture extends PIXI.Texture {
         
     }
 }